Add meal choices to RSVP form.
Also remove mailing address and shuttle option. Carry those forward but don't
allow for these to be edited.

Currently, the meal choices are shown for all guests. Need to remove the
meal choices for guests which decline dynamically.

If a guest declines but chooses a meal choice, that is not saved.